fix(react/runtime): make `setupLynxEnv` run earlier
