+++ .changeset/funny-jokes-march.md
@@ -1,0 +1,5 @@
+---
+"@lynx-js/react": patch
+---
+
+Fix 'SystemInfo is not defined' error when using MTS and not importing anything manually from the react package.
